Fix for matlab tests which appear to have bodies without parents
This test confirms that only a single body can have no parents, the
world body.  The previous used the body name to determine if the body
was the world body.  Matlab code changes that name as part of its
optimizations, invalidating this test.  However, in both cases, the
body index should be common between the two.