Add automatic header installation
Add `guess_hdrs` option to `install`, to automatically install
transitive headers of a library, with options to restrict to headers in
the same package or same workspace. Add install rule for libdrake.so.

This provides an easy mechanism to guess what headers need to be
installed that is consistent with how `drake_header_tar` computes
headers (except that we exclude external headers in the expectation that
those will have their own install rules), with the trade-off that it
doesn't differentiate between public and private headers (also
consistent with `drake_header_tar`).