Allow customizing more install destinations
Add ability to override the default install destinations for archive,
library and runtime artifacts. This is needed for proper installation of
compiled Python modules. (The only apparent way to install these to the
correct location is for the user to override these destinations, as
there is no obvious way to differentiate between a Python module and any
other compiled library.)