Fix LCM, Drake install
Switch install destination for LCM's and Drake's CMake files to use the
equivalent of the CPS path. (CMake supports both `cmake/pkg` and
`pkg/cmake`, but CPS specifies `cps/pkg`; LCM upstream and the logic
prior to this commit uses `pkg/cmake`.) Fix unqualified name in
"Default-Components" which ought to be an error (cps2cmake accepts it,
even though technically it shouldn't, but produces a _LIBRARIES with the
unqualified name as a result, which won't be usable as the actual
imported targets are qualified).

fixup locations