Added horrible hack for query
In order to invoke queries on geometry_system, geometry system has to be
able to update its state from input ports. Because this isn't part of the
output port, this must be done in some other way. Currently, I do a
brute force solution where I jam the values in a const_casted state.

a) I need a better solution.
b) I need to refactor this so it can be handled appropriately.