[PATCH] Adds a profiler to drake
This is a simple patch for adding in the profiler -- I can cherry-pick this
commit to do some profiling and then trash it. To use it:

In main function:
```c++
#include "drake/common/profiler.h"

int main() {
...

std::cout << TableOfAverages() << "\n";
return 0;
}
```

In profiling code (which can include the main program):

```c++
#include "drake/common/profiler.h"

void Foo() {
   static const TimerIndex foo_timer = addTimer("TimerForFoo");
   ... // code not included in profiling
   startTimer(foo_timer);
   ... // code to profile.
   lapTimer(foo_timer);
   ... // code not included in profiling
}

In building and running with profiler enabled:

bazel run --copt -DENABLE_TIMERS build/target:thing

TODO: Fix the function names to be drake compliant.