Brings "bad sqrt" unit tests written by Sean in #14940.
If the relative velocity of the two bodies at a quadrature point is zero
(e.g., if the bodies are both stationary), the calculation of traction,
which included a square root in calculation of the norm of the relative
velocity, would introduce NaNs in the derivatives for AutoDiffXd-valued
contact forces.

This cause for this is that, more or less, the derivative of sqrt is
1/sqrt. If I'm taking the square root of zero, I get 0/0 in the
derivatives. This eliminates the hard norm and replaces it with the soft
norm already computed.